+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+### Added
+* Panel: New component ([#2711](https://github.com/dso-toolkit/dso-toolkit/issues/2711))
+
 ### Tasks
 * Build: Release automatisering ([#1337](https://github.com/dso-toolkit/dso-toolkit/issues/1337))